hasql: derive ENUM/CITEXT decoders, widen INT/NUMERIC columns
WrappedEnumSC now takes a leading `Maybe Symbol` pinning the backing PG
type. `'Nothing` keeps the old text-backed behaviour; `('Just "schema.name")`
swaps in `D.enum`/`E.enum` so hasql-interpolate's strict OID lookup resolves
real `CREATE TYPE … AS ENUM` columns through generic DecodeRow without
per-query `::text` casts. `HI.DecodeValue (CI Text)` switches to `D.citext`
to cover the citext extension type.

Updated qualified sites: Permissions, IssueType, AnomalyTypes,
AnomalyActions, QueryLibType, KeyKind. Everything else gets `'Nothing`.

Migrations:
- 0097 drops the `email` domain (now plain citext + format CHECK) and
  widens `apis.log_patterns.baseline_samples` /
  `projects.replay_sessions.event_file_count` to BIGINT.
- 0098 batches the rest of the int4 → int8 widening (issues affected_*,
  llm_enhancement_version, seq_num; error_patterns occurrences/etc.;
  query_monitors check_interval_mins/threshold/notification_count/…;
  notification_rate_limit.count) plus `apis.issues.error_rate` → float8.

Code fixes surfaced by the strict decoders:
- `Issue` gains `cooldownUntil` / `lastNotifiedAt` (rows existed since
  0075/0085) and the IssueL list SQL selects them.
- BackgroundJobs SafetyNet uses `Telemetry.otelSpanColsSql` (was missing
  `errors` + `message_size_bytes`).
- consumeNotificationToken returns `count` directly (no int4 cast).
- make_interval call casts hours to `int` (only signature PG provides).
- Anomalies VM format_examples now `TEXT[]` (matches Haskell Vector Text).
- LogQueries bis/cnts and totalSessions widened to Int64.
- Parser alert select emits `count(*)::float8` for Double consumers.
- Projects.downgradeToFree / upgradeToPaid cast bigint params to text for
  the TEXT `order_id` comparison.
- isInCooldown SELECTs `1::bigint` so the Int64 decoder matches.

+++ src/Models/Apis/LogQueries.hs
@@ -460,10 +460,10 @@ data RawSessionRow = RawSessionRow
   , lastSeen :: UTCTime
   , durationNs :: Int64
   , traceCount :: Int64
-  , totalSessions :: Int32
+  , totalSessions :: Int64
   , services :: V.Vector Text
-  , bis :: V.Vector Int32
-  , cnts :: V.Vector Int32
+  , bis :: V.Vector Int64
+  , cnts :: V.Vector Int64
   , landingUrl :: Maybe Text
   , userAgent :: Maybe Text
   , firstError :: Maybe Text
